+++ tinylink/__init__.py
@@ -1,4 +1,4 @@
 from .consts import *  # noqa
-from .link import Frame, TinyLink  # noqa
+from .link import AsyncTinyLink, Frame, TinyLink  # noqa
 
 __version__ = "2.0.0"